ArticleShip/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
duckduckgo-search>=5.2.2
google-genai
python-dotenv>=1.0.1
pydantic>=2.6.1
pymongo>=4.7.3